Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

Please stick with package rather than yum as it is already done in the module. (Sole reason for the -1)

Extra: Since this is not part of the 'regular' workflow, and this is being addressed by the cert team, it would be nice to have this two tasks in its own yaml file under a patches directory (ie. patches/rhosp-release-TICKETNUMBER) that we could include in the mainl.yml, this way we 'know' what's the real workflow and what is a patch we apply to bypass current situation.

As I say, its extra, I'd +2 if s/yum/package is done

Patch-set: 3
Label: Code-Review=-1