add the tests directory
The purpose of this script is to give us a way to quickly validate the
role behave has expected with different OSP releases.

It uses Docker, because we depend on a RHEL7 environment

The role depends on some RH interne resources. I believe it's acceptable
since the role is also RH centric.

Change-Id: Iafa5b364b203ca046f7d84aa6c3cd385d2cb49d9