Update patch set 5
Patch Set 5: Code-Review-1

I agree with @dsavinea. Docker - in that case - doesn't bring any added value.   
Actually it makes the patch bigger than what it could be.                           
                                                                                    
At the zuul level, we can ensure the CI job is run from a RHEL7 image;              
removing the need for docker.                                                       
                                                                                    
Since it is run on a RHEL image, we can rely directly on rhsm to subscribe to       
the proper channel and install the proper packages; removing the need for all       
those repo files. (Plus remember that ansible-role-openstack-certification is a   
public project - we are leaking URLS here). This also allows for the test           
playbook to be way smaller than what it currently is.                               
                                                                                    
Summing it up, specifying that the CI job should be run on a RHEL7 would remove  
the following files from this patchset:                                             
                                                                                    
 * tests/files/overcloudrc                                                          
 * tests/files/rhcert.repo                                                          
 * tests/files/rhel.repo                                                            
 * tests/files/rhos_osp10.repo                                                      
 * tests/files/rhos_osp12.repo                                                      
 * tests/files/rhos_osp13.repo                                                      
 * tests/inventory                                                                  
 * tests/requirements.yml                                                           
 * tests/roles/openstack-certification                                              
                                                                                    
And causes the test.yml file to focus just on what it should be doing:              
                                                                                    
  * Run the certification test suite with proper params                             
  * Ensure we have a result back (xml file with content)

Patch-set: 5
Reviewer: Gerrit User 115 <115@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=-1