Update patch set 5
Patch Set 5:

I think we are not totally on the same page. As a developer, I want to be able to quickly validate my last changeset. I don't run RHEL and I have access to the RH resources. This is the case I was focusing on.

The Zuul integration is another thing, if the VM is outside of the network.

Patch-set: 5
Reviewer: Gerrit User 30 <30@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=0