Update patch set 5
Patch Set 5:

Let's clarify the goal of this patchset: Be able to run a test suite that will give me feedback about my current change. Is that correct?

Now, they are two audiences:

  * CI/Gating system
  * Developer on her laptop

If the patchset is updated with what was said in the previous comment, both audiences are adressed with a common code base.

  * As a CI system, I will git clone from a RHEL machine this repo and run the test (no hard dependency)

  * As a developer, I already have docker installed with a RHEL image, I can run the test

Code and files remains as focus as possible to its core, installing and running openstack-certification in a platform independent way. Wether one prefers to run the test in a VM, in a container, or any other way, this is independant from the project code and hence should remain outside of the scope of the project.

Patch-set: 5