Openstack and Baremetal packages are exclusive
When installating both openstack and baremetal rpm packages on the
same host, we will have supportable and sosreport tests ran twice
because both package are exclusive (at run time).
When baremetal is present in one of the certification types we don't
need to executed the openstack tasks (test_config.json file). The
opposite is also true.

Change-Id: I81b8b4ed23ecf854c2b112892c2000208604a3ce