Update patch set 1
Patch Set 1:

> Not sure this is the right way to do multi host.
 > An ansible role should not have specific code related to a
 > host/group (ex: openstack_certification_tests_{compute,controller}
 > doesn't make sense). This needs to be done at playbook level.

Maybe, but how else should it be done when we need to run different test sets on each node? In my opinion it makes perfect sense, because me as a user who wants to run certification tests, I just want to run the role and let it to deal with choosing the right test set for the node it's being run on. So the only thing a user needs to do is to set hosts file and execute a very simple playbook which runs the role (like the one in this review). 

Also by this approach, the tests are executed on each node in parallel. If I created more playbooks or run the role multiple times with different test set and on a different role, the tests would be run sequentially.

But I'm gonna check if it's possible to run playbooks or roles in parallel.

Patch-set: 1