Update patch set 1
Patch Set 1:

$ cat hosts
[controller]
heat-admin@192.168.24.13

[compute]
heat-admin@192.168.24.7

[controller:vars]
test=controller

[compute:vars]
test=compute

$ cat playbook.yaml
- hosts: all
  tasks:
    - debug:
        var: test

    - debug:
        msg: "testing"

Good, that ^^ would work as we want, the tasks were executed in parallel and each of the nodes had the correct value of test variable.
What I personally don't like about this approach is that the variables are defined in hosts, it's just not the first place where I would take a look at.
Another problem is, that this role is also an infrared plugin. When running the role from infrared, we can't define group_vars, so we would need to call the plugin three times, which is slow and it creates more complications in general. That's why I would prefer to have the logic inside the role. It's easily manageable for everyone.

Patch-set: 1